Add scroll listener, minor cleanup
