+++ source/background/index.js
@@ -4,7 +4,7 @@ import log from "../shared/library/log.js";
 import { attachBrowserStateListeners } from "./library/browserEvents.js";
 import { updateContextMenu } from "./library/contextMenu.js";
 import { getBrowser } from "../shared/library/browser.js";
-import { checkUnlockPossibility, watchForSourcesAutoLock } from "./library/archives.js";
+import { checkUnlockPossibility, watchForSourcesAutoLock } from "./library/autoLock.js";
 import { watchStorage as watchStorageForConfig } from "./library/config.js";
 import store from "./redux/index.js";
