+++ source/shared/library/mouseEvents.js
@@ -3,10 +3,16 @@ import { trackUserActivity } from "./messaging";
 
 export function trackMouseMovement() {
     const debouncedUserActivity = debounce(() => trackUserActivity(), 250, /* trailing: */ false);
-
-    let handleMousemove = () => {
+    const handleMousemove = () => {
         debouncedUserActivity();
     };
-
     document.addEventListener("mousemove", handleMousemove, { passive: true });
+}
+
+export function trackScrolling() {
+    const debouncedUserActivity = debounce(() => trackUserActivity(), 250, /* trailing: */ false);
+    const handleScrollUpdate = () => {
+        debouncedUserActivity();
+    };
+    window.addEventListener("scroll", handleScrollUpdate);
 }
